Rename bot to "ComposerHostBotDotNet"
