+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/.gitignore
