+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/ComposerHostBotDotNet.csproj
