+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/Nuget.config
