+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/Properties/launchSettings.json
