+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/Startup.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace WaterfallHostBotComposer
+namespace ComposerHostBotDotNet
 {
     public class Startup
     {
