+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/dialogs/SsoDialog/recognizers/SsoDialog.lu.dialog
