+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/dialogs/SsoSignInDialog/SsoSignInDialog.dialog
