+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/dialogs/SsoSignInDialog/recognizers/SsoSignInDialog.lu.dialog
