+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/media/create-azure-resource-command-line.png
