+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/media/publish-az-login.png
