+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/recognizers/ComposerHostBotDotNet.en-us.lu.dialog
@@ -1,7 +1,7 @@
 {
   "$kind": "Microsoft.OrchestratorRecognizer",
   "modelFolder": "=settings.orchestrator.models.en",
-  "snapshotFile": "=settings.orchestrator.snapshots.WaterfallHostBotComposer_en_us",
+  "snapshotFile": "=settings.orchestrator.snapshots.ComposerHostBotDotNet_en_us",
   "detectAmbiguousIntents": true,
   "disambiguationScoreThreshold": 0.05
 }
