+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/recognizers/ComposerHostBotDotNet.lu.dialog
@@ -1,5 +1,5 @@
 {
   "$kind": "Microsoft.MultiLanguageRecognizer",
-  "id": "ORCHESTRATOR_WaterfallHostBotComposer",
+  "id": "ORCHESTRATOR_ComposerHostBotDotNet",
   "recognizers": {}
 }
