+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/schemas/sdk.schema
