+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/schemas/update-schema.ps1
