+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/schemas/update-schema.sh
