+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/settings/appsettings.json
@@ -79,7 +79,7 @@
     "defaultLanguage": "en-us",
     "endpoint": "",
     "environment": "composer",
-    "name": "WaterfallHostBotComposer"
+    "name": "ComposerHostBotDotNet"
   },
   "publishTargets": [],
   "qna": {
@@ -88,7 +88,7 @@
     "qnaRegion": "westus"
   },
   "runtime": {
-    "command": "dotnet run --project WaterfallHostBotComposer.csproj",
+    "command": "dotnet run --project ComposerHostBotDotNet.csproj",
     "customRuntime": true,
     "key": "adaptive-runtime-dotnet-webapp",
     "path": "../"
