+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/echoskillbotdotnet-manifest.json
