+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/echoskillbotdotnet21-manifest.json
