+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/echoskillbotdotnetv3-manifest.json
