+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/echoskillbotjs-manifest.json
