+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/echoskillbotjsv3-manifest.json
