+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/echoskillbotpython-manifest.json
