+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/teamsskillbotdotnet-manifest.json
