+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/teamsskillbotjs-manifest.json
