+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/teamsskillbotpython-manifest.json
