+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/waterfallskillbotdotnet-manifest.json
