+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/waterfallskillbotjs-manifest.json
