+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/skills/waterfallskillbotpython-manifest.json
