+++ Bots/DotNet/Consumers/Composer/ComposerHostBotDotNet/wwwroot/default.htm
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>WaterfallHostBotComposer</title>
+    <title>ComposerHostBotDotNet</title>
     <style>
         body {
             margin: 0px;
@@ -338,7 +338,7 @@
     <header class="header">
         <div class="header-inner-container">
             <div class="header-icon" style="display: inline-block"></div>
-            <div class="header-text" style="display: inline-block">WaterfallHostBotComposer</div>
+            <div class="header-text" style="display: inline-block">ComposerHostBotDotNet</div>
         </div>
     </header>
     <div class="row">
