+++ Bots/DotNet/FunctionalTestsBots.sln
@@ -39,7 +39,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleHostBotComposer", "Co
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EchoSkillBotComposer", "Skills\Composer\EchoSkillBotComposer\EchoSkillBotComposer.csproj", "{9BA47CF9-7D90-4B5C-A9FA-01797A435D53}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WaterfallHostBotComposer", "Consumers\Composer\WaterfallHostBotComposer\WaterfallHostBotComposer.csproj", "{80AF5608-8126-439F-B463-197E6A17D941}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ComposerHostBotDotNet", "Consumers\Composer\ComposerHostBotDotNet\ComposerHostBotDotNet.csproj", "{E3FCDDD4-553D-44BA-8B2F-82BA778261BC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -83,10 +83,10 @@ Global
 		{9BA47CF9-7D90-4B5C-A9FA-01797A435D53}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9BA47CF9-7D90-4B5C-A9FA-01797A435D53}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9BA47CF9-7D90-4B5C-A9FA-01797A435D53}.Release|Any CPU.Build.0 = Release|Any CPU
-		{80AF5608-8126-439F-B463-197E6A17D941}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{80AF5608-8126-439F-B463-197E6A17D941}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{80AF5608-8126-439F-B463-197E6A17D941}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{80AF5608-8126-439F-B463-197E6A17D941}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E3FCDDD4-553D-44BA-8B2F-82BA778261BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E3FCDDD4-553D-44BA-8B2F-82BA778261BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E3FCDDD4-553D-44BA-8B2F-82BA778261BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E3FCDDD4-553D-44BA-8B2F-82BA778261BC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -105,7 +105,7 @@ Global
 		{41BC9547-FD9E-40E1-B1D9-C3F25BC9A2F3} = {6EF38D8C-7953-4D54-BA31-FFF055B7B217}
 		{FDC53B3A-0E15-4FDF-A587-05C8F90BC2B6} = {33FA275E-9E5D-4582-BFF7-24B2C5DB2962}
 		{9BA47CF9-7D90-4B5C-A9FA-01797A435D53} = {33E2281F-C6A7-40EC-961A-FF9FF254FDDC}
-		{80AF5608-8126-439F-B463-197E6A17D941} = {33FA275E-9E5D-4582-BFF7-24B2C5DB2962}
+		{E3FCDDD4-553D-44BA-8B2F-82BA778261BC} = {33FA275E-9E5D-4582-BFF7-24B2C5DB2962}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2B77A921-EEA6-4006-ABD2-159C92F9F874}
